backend/requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# startup.py: 8
Flask == 1.1.1

# forms.py: 5
WTForms == 2.2.1

# forms.py: 1
WTForms_JSON == 0.3.3

# startup.py: 2,3
pydub == 0.23.1

# forms.py: 3
python_dateutil == 2.8.1

# startup.py: 21
pytz == 2019.3

# startup.py: 13
requests == 2.22.0

# startup.py: 10
tzlocal == 2.0.0
